refactor: reorganize lib/ by domain
Group lib/ files into domain folders:
- lib/inquiry/ — schema, submit, rate-limiter, email/
- lib/services/ — types and data split into separate files
- lib/cosmos/ — simulation logic

Update all import paths throughout the codebase.
