+++ lib/cosmos/simulation.ts
