+++ lib/services/types.ts
@@ -1,0 +1,18 @@
+import type { StaticImageData } from 'next/image';
+
+export type Service = {
+  slug: string;
+  title: string;
+  shortDescription: string;
+  description: string;
+  longDescription: string;
+  features: string[];
+  process: [string, string, string][];
+  deliverables: string[];
+  pricingNote: string;
+  timeNote: string;
+  designation: string;
+  glyph: string;
+  image: StaticImageData;
+  imageAlt: string;
+};
